feat: use `u32` + `Number.MAX_SAFE_INTEGER`

+++ crates/rspack_binding_values/src/raw_options/raw_module/mod.rs
@@ -435,7 +435,7 @@ impl From<RawCssModuleParserOptions> for CssModuleParserOptions {
 #[derive(Debug, Default)]
 #[napi(object)]
 pub struct RawJsonParserOptions {
-  pub exports_depth: Option<f64>,
+  pub exports_depth: Option<u32>,
 }
 
 impl From<RawJsonParserOptions> for JsonParserOptions {
